requirements.txt
maplotlib
numpy
opencv-python
Pillow
pandas
scikit-learn
streamlit